saved comics dont display on rest of site

+++ client/app/the-feed/the-feed.controller.ts
@@ -3,7 +3,7 @@
 class TheFeedCtrl {
   private comicsToShowDescription: string[];
 
-  constructor($scope, $http, Auth, comic){
+  constructor($scope, $http, Auth, comic, socket){
     this.isContributor = Auth.isUser;
     this.Auth = Auth;
     this.$scope = $scope;
@@ -15,34 +15,28 @@ class TheFeedCtrl {
     // Grab the initial set of available comics
     $http.get('/api/Comics').success(function(Comics) {
       $scope.Comics = Comics;
+      socket.syncUpdates('Comics', Comics);
       if (Comics[0] == undefined){
         $scope.noComics = true;
       }
       $scope.loadingComics = false;
     });
 
-/*    comic.getComics().success(function(Comics) {
-      $scope.Comics = Comics;
-      if (Comics[0] == undefined){
-        $scope.noComics = true;
-      }
-      $scope.loadingComics = false;
-    });*/
-
   }
 
   editComic(Comic){
-    //console.log(Comic._id);
-    var stringVal = "blenhwhke";
-    //this.$http.get('api/Comics/' + Comic._id).success(function(data){editObj = data});
     this.comic.setComic(Comic._id);
-    this.comic.setString(stringVal);
+    console.log(this.Auth.getCurrentUser()._id);
+    this.$http.post('/api/Comics/' + Comic._id + '/contributors' ,{contributors: this.Auth.getCurrentUser()._id, notSaved: false});
+    this.$http.post('api/users/' + this.Auth.getCurrentUser()._id + '/my-comics', {id: this.Auth.getCurrentUser()._id, myComics: Comic._id});
     //window.location.href='/create-a-comic';
   }
 
   deleteComic(Comic) {
+    if (Comic.notSaved == true){
       this.$http.delete('/api/Comics/' + Comic._id);
-      window.location.href='/the-feed';
+      window.location.href='/the-feed';}
+    else {console.log('you done fucked up')}
   };
   addToShowDescription(comicId: string){
     this.comicsToShowDescription.push(comicId);
